refactor: LottoResult 클래스의 메서드들을 Lotto 클래스로 이동
- 과도한 분리를 방지하고자 LottoResult 클래스 삭제하고 LottoResult의 메서드들을 Lotto 클래스로 이동
