+++ docs/README.md
@@ -59,7 +59,13 @@
 
 * number(당첨 번호)가 유효하면 외부에서 접근하지 못하도록 private 접근자가 있는 #number변수에 number를 할당
 
-* 외부에서 사용할 수 있도록 각 번호를 숫자로 타입변환한 당첨 번호 배열을 반환하는 기능(getWinningNumber)
+- [x] 외부에서 사용할 수 있도록 각 번호를 숫자로 타입변환한 당첨 번호 배열을 반환하는 기능(getWinningNumber)
+
+1등 - (6개), 2등 - (5개 + 보너스), 3등 - (5개), 4등 - (4개), 5등 - (3개)
+
+- [x] 사용자가 발행한 모든 로또 번호들을 당첨 번호 6자리와 비교하여 일치하는 개수를 새로운 배열로 저장하는 기능 
+
+- [x] 일치 개수가 3, 4, 6 이면 당첨 내역의 개수의 증가를 반복하는 기능 (보너스 번호를 검사해야 하므로 5개가 일치하는 경우만 제외하고 증가시킴) 
 
 ### 5. 보너스 번호 클래스(Bonus)
 
@@ -70,17 +76,8 @@
 
 - [x] 숫자 5개가 일치하는 로또 목록을 순회하며 보너스 번호와의 일치 여부 확인하고 보너스 번호가 일치하면 2등, 불일치하면 3등의 당첨 개수를 증가시키는 기능
 
-### 6. 당첨 결과 관련 클래스(LottoResult)
 
-1등 - (6개), 2등 - (5개 + 보너스), 3등 - (5개), 4등 - (4개), 5등 - (3개)
-
-- [x] 사용자가 발행한 모든 로또 번호들을 당첨 번호 6자리와 비교하여 일치하는 개수를 새로운 배열로 저장하는 기능 
-
-- [x] 일치 개수가 3, 4, 6 이면 당첨 내역의 개수의 증가를 반복하는 기능 (보너스 번호를 검사해야 하므로 5개가 일치하는 경우만 제외하고 증가시킴) 
-
-- [x] 입력받은 액수와 총 당첨금을 토대로 수익률 계산하여 반환하는 기능
-
-### 7. 상수 객체 파일 (constants)
+### 6. 상수 객체 파일 (constants)
 
 - [x] 입력 요청 메세지 객체 (INPUT_QUERY)
