+++ src/Lotto.js
@@ -1,4 +1,5 @@
-const { checkWinningNumber } = require("./Validation")
+const { prizeCount } = require("../constants");
+const { checkWinningNumber } = require("../Validation")
 
 class Lotto {
   #numbers;
@@ -10,6 +11,36 @@ class Lotto {
 
   getWinningNumber() {
     return this.#numbers.split(",").map((item) => Number(item));
+  }
+
+  checkEachLottoNumber(publishedLotto) {
+    const matchedNumberList = [];
+    const winningNumber = this.getWinningNumber();
+
+    publishedLotto.forEach((eachUserLottoNumber) => {
+      matchedNumberList.push(this.getMatchNumberList(winningNumber, eachUserLottoNumber).length);
+    });
+    this.countMatchNumber(matchedNumberList)
+  }
+
+  getMatchNumberList(winningNumber, eachUserLottoNumber) {
+    return winningNumber.filter((number) => eachUserLottoNumber.includes(number));
+  }
+
+  countMatchNumber(matchedNumberList) {
+    matchedNumberList.forEach((number) => {
+      switch (number) {
+        case 6:
+          prizeCount.first += 1;
+          break;
+        case 4:
+          prizeCount.fourth += 1;
+          break;
+        case 3:
+          prizeCount.fifth += 1;
+          break;
+      }
+    });
   }
 
 }
